Real-time render

+++ src/hsv.cpp
@@ -1,7 +1,7 @@
 #include "hsv.h"
 
 HSV::HSV(){}
-HSV::HSV(int hue, int saturation, int value) : _hue(hue), _saturation(saturation), _value(value){}
+HSV::HSV(float hue, float saturation, float value) : _hue(hue), _saturation(saturation), _value(value){}
 
 float HSV::getHue(){
     return _hue;
